+++ R/GibbsBvsF.R
@@ -23,10 +23,11 @@
 #' The options \code{prior.models="SBSB"}, \code{prior.models="ConstConst"} and \code{prior.models="SBConst"}
 #' acknowledge the "grouped" nature of the dummy variables representing
 #' factors through the use of two stage
-#' priors described in Garcia-Donato and Paulo (2018). In the first stage probabilities over factors and numerical
+#' priors described in Garcia-Donato and Paulo (2021). In the first stage probabilities over factors and numerical
 #' variables are specified and (conditional on these) within the second stage
 #' the probablities are apportioned over the different submodels defined
-#' by the dummies. The default option is "SBSB" which uses in both stages an assignment
+#' by the dummies. The default (and recommended, for the reasons argued in Garcia-Donato and Paulo,2021) 
+#' option is "SBSB" which uses in both stages an assignment
 #' of the type Scott-Berger so inversely proportional to the number of models of the same dimension. The
 #' option "ConstConst" implements a uniform prior for both stages while "SBConst" uses a Scott-Berger prior
 #' in the first stage and it is uniform in the second stage. Within all these priors, the prior inclusion probabilities
@@ -46,9 +47,9 @@
 #' contains the intercept and by default, the null model is defined
 #' to be the one with just the intercept
 #' @param prior.betas Prior distribution for regression parameters within each
-#' model. Possible choices include "Robust", "Liangetal", "gZellner",
-#' and "ZellnerSiow" (see details in \code{\link[BayesVarSel]{Bvs}}).
-#' @param prior.models Prior distribution over the model space. Possible
+#' model (to be literally specified). Possible choices include "Robust", "Robust.G", "Liangetal", "gZellner",
+#' "ZellnerSiow" (see details in \code{\link[BayesVarSel]{Bvs}}).
+#' @param prior.models Prior distribution over the model space (to be literally specified). Possible
 #' choices (see details) are "Const", "SB", "ConstConst", "SBConst" and "SBSB" (the default).
 #' @param n.iter The total number of iterations performed after the burn in
 #' process.
@@ -98,6 +99,8 @@
 #' \item{positions }{A binary matrix with \code{p} rows and (pnum+sum_j L_j) columns. The 1's identify, for each variable (row) the position (column)
 #' of dummies (in case of factor) or of the numerical variable grouped on that variable. (Its use is conceived for internal purposes).}
 #' \item{positionsx }{A \code{p} dimensional binary vector, stating which of the competing variables is a numerical variable. (Its use is conceived for internal purposes).}
+#' \item{prior.betas}{prior.betas}
+#' \item{prior.models}{prior.models}
 #' \item{method }{\code{gibbsWithFactors}}
 #' @author Gonzalo Garcia-Donato and Anabel Forte
 #' @seealso \code{\link[BayesVarSel]{plot.Bvs}} for several plots of the result.
@@ -112,8 +115,8 @@
 #' variable selection problems with large model spaces. Journal of the American
 #' Statistical Association, 108: 340-352.
 #'
-#' Garcia-Donato, G. and Paulo, R. (2018) Including factors in Bayesian variable selection
-#' problems. arXiv:1709.07238.
+#' Garcia-Donato, G. and Paulo, R. (2021) Variable selection in the presence of factors: 
+#' a model selection perspective. Journal of American Statistical Association, Ahead-of- print(1-11).
 #'
 #' George E. and McCulloch R. (1997) Approaches for Bayesian variable
 #' selection. Statistica Sinica, 7, 339:372.
@@ -381,13 +384,13 @@ GibbsBvsF <-
     if (prior.betas == "Liangetal"){write(4, ncolumns=1, file=paste(wd, "/typeofBF.txt", sep = ""))}
     if (prior.betas == "gZellner"){write(2, ncolumns=1, file=paste(wd, "/typeofBF.txt", sep = ""))}
     if (prior.betas == "ZellnerSiow"){write(5, ncolumns=1, file=paste(wd, "/typeofBF.txt", sep = ""))}
-    if (prior.betas == "Robust2"){write(6, ncolumns=1, file=paste(wd, "/typeofBF.txt", sep = ""))}
+    if (prior.betas == "Robust.G"){write(6, ncolumns=1, file=paste(wd, "/typeofBF.txt", sep = ""))}
 
 	  if (prior.betas == "FLS"){stop("Prior FLS not yet supported\n")}
 
 		if (prior.betas != "Unitary" & prior.betas != "Robust" & prior.betas != "Liangetal" &
 			    prior.betas != "gZellner" & prior.betas != "ZellnerSiow" & prior.betas != "FLS" &
-					prior.betas != "Robust2") {stop("Dont recognize the prior for betas\n")}
+					prior.betas != "Robust.G") {stop("Dont recognize the prior for betas\n")}
 
 		if (prior.models=="SBSB"){method<- "rSBSB"}
 		if (prior.models=="ConstConst"){method<- "rConstConst"}
@@ -562,6 +565,9 @@ GibbsBvsF <-
     result$call <- match.call()
 
     result$method <- "gibbsWithFactors"
+	result$prior.betas<- prior.betas	
+	result$prior.models<- prior.models
+	
     class(result)<- "Bvs"
     result